[IMP] Merged code of pos.config into one file, added migration script to fix the issue of Cheques Journals in pos.config
