+++ coop_point_of_sale/models/__init__.py
@@ -3,14 +3,12 @@
 # @author: La Louve
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-# from . import res_config
 from . import pos_config
 from . import product_product
 from . import product_template
 from . import pos_order_line
 from . import pos_session
 from . import pos_order
-from . import pos_config_settings
 from . import res_company
 from . import pos_category
 from . import stock_picking
